+++ BotDofus/src/Protocol/Messages/Data/game/guild/tax/GuildFightTakePlaceRequestMessage.h
@@ -1,9 +1,9 @@
 #ifndef GUILDFIGHTTAKEPLACEREQUESTMESSAGE_H
 #define GUILDFIGHTTAKEPLACEREQUESTMESSAGE_H
 
-#include "src/Engines/IO/Network/Utils/FuncTree.h"
-#include "src/Protocol/Messages/AbstractMessage.h"
-#include "src/Protocol/Messages/Data/game/guild/tax/GuildFightJoinRequestMessage.h"
+#include "src/engines/io/network/utils/FuncTree.h"
+#include "src/protocol/messages/AbstractMessage.h"
+#include "src/protocol/messages/data/game/guild/tax/GuildFightJoinRequestMessage.h"
 
 class GuildFightTakePlaceRequestMessage : public GuildFightJoinRequestMessage
 {
@@ -16,7 +16,7 @@ public:
   void deserializeAsyncAs_GuildFightTakePlaceRequestMessage(FuncTree tree);
   GuildFightTakePlaceRequestMessage();
 
-  int replacedCharacterId;
+  double replacedCharacterId;
 
 private:
   void _replacedCharacterIdFunc(Reader *input);
